move the transforms in a strip-test-selectors workspace
