+++ strip-test-selectors/.eslintrc.js
@@ -1,0 +1,12 @@
+module.exports = {
+  root: true,
+  extends: ['simplabs', 'plugin:prettier/recommended'],
+  parserOptions: {
+    sourceType: 'script',
+    ecmaVersion: 2015,
+  },
+  env: {
+    browser: false,
+    node: true,
+  },
+};
