+++ strip-test-selectors/.gitignore
@@ -1,0 +1,26 @@
+# See https://help.github.com/ignore-files/ for more about ignoring files.
+
+# compiled output
+/dist
+/tmp
+*.tgz
+
+# dependencies
+/node_modules
+/bower_components
+/node-tests/multidep_modules
+
+# misc
+/.eslintcache
+/.sass-cache
+/connect.lock
+/coverage/*
+/libpeerconnection.log
+npm-debug.log*
+yarn-error.log
+testem.log
+
+# ember-try
+.node_modules.ember-try/
+bower.json.ember-try
+package.json.ember-try
