+++ strip-test-selectors/.prettierrc.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  arrowParens: 'avoid',
+  printWidth: 100,
+  singleQuote: true,
+  trailingComma: 'es5',
+};
