+++ strip-test-selectors/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "strip-test-selectors",
+  "version": "0.0.0",
+  "description": "Enabling better Test selectors in Ember.js applications.",
+  "license": "MIT",
+  "author": "Mainmatter GmbH",
+  "repository": "https://github.com/mainmatter/ember-test-selectors",
+  "files": [
+    "src"
+  ],
+  "scripts": {
+    "lint": "eslint . --cache"
+  },
+  "devDependencies": {
+    "eslint": "8.57.1",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-config-simplabs": "0.4.0",
+    "eslint-plugin-prettier": "4.2.1",
+    "prettier": "2.8.8"
+  }
+}
