+++ strip-test-selectors/src/strip-data-test-properties-plugin6.js
