+++ strip-test-selectors/src/strip-test-selectors.js
