+++ tests/node-tests/strip-data-test-properties-plugin-test.js
@@ -3,7 +3,7 @@ let assert = require('assert');
 
 let babel7 = require('@babel/core');
 
-let StripDataTestPropertiesPlugin6 = require('ember-test-selectors/strip-data-test-properties-plugin6');
+let StripDataTestPropertiesPlugin6 = require('strip-test-selectors/src/strip-data-test-properties-plugin6');
 
 function testFixture(name) {
   it(`fixture: ${name}`, function () {
